+++ lib/components/Grid/constants.ts
@@ -1,0 +1,6 @@
+export const defaultGridProps = {
+  vertical: false,
+  horizontal: true,
+  strokeDasharray: "3",
+  className: "stroke-1",
+};
