+++ lib/components/Grid/index.tsx
@@ -1,11 +1,3 @@
 import { CartesianGrid as Grid } from 'recharts';
 
-export const defaultGridProps = {
-  vertical: false,
-  horizontal: true,
-  strokeDasharray: "3",
-  className: "stroke-1",
-};
-
-
 export { Grid };
