+++ lib/components/XAxis/constants.ts
@@ -1,0 +1,9 @@
+export const defaultXAxisProps = {
+  paddingValue: 0,
+  startEndOnly: false,
+  intervalType: 'preserveStartEnd',
+  className: "text-xs fill-gray-600",
+  tickGap: 5,
+  fill: "",
+  stroke: "",
+};
