+++ lib/components/XAxis/index.tsx
@@ -1,16 +1,5 @@
 import { XAxisProps } from 'recharts';
 
-export const defaultXAxisProps = {
-  paddingValue: 0,
-  startEndOnly: false,
-  intervalType: 'preserveStartEnd',
-  className: "text-xs fill-gray-600",
-  tickGap: 5,
-  fill: "",
-  stroke: "",
-};
-
-
 export { XAxis } from 'recharts';
 
 export type { XAxisProps };
