+++ lib/components/YAxis/constants.ts
@@ -1,0 +1,7 @@
+export const defaultYAxisProps = {
+  axisLine: false,
+  tickLine: false,
+  className: "text-xs fill-gray-600",
+  fill: "",
+  stroke: "",
+};
