+++ lib/components/YAxis/index.tsx
@@ -1,13 +1,5 @@
 import { YAxisProps } from 'recharts';
 
-export const defaultYAxisProps = {
-  axisLine: false,
-  tickLine: false,
-  className: "text-xs fill-gray-600",
-  fill: "",
-  stroke: "",
-};
-
 export { YAxis } from 'recharts';
 
 export type { YAxisProps };
