+++ lib/main.ts
@@ -4,11 +4,14 @@ export { BarChart } from './components/BarChart'
 export type { BarChartProps } from './components/BarChart'
 export { BarItem } from './components/BarChart/BarItem'
 export type { BarItemProps } from './components/BarChart/BarItem'
-export { XAxis, defaultXAxisProps } from './components/XAxis'
+export { XAxis } from './components/XAxis'
+export { defaultXAxisProps } from './components/XAxis/constants'
 export type { XAxisProps } from './components/XAxis'
-export { YAxis, defaultYAxisProps } from './components/YAxis'
+export { YAxis } from './components/YAxis'
+export { defaultYAxisProps } from './components/YAxis/constants'
 export type { YAxisProps } from './components/YAxis'
-export { Grid, defaultGridProps } from './components/Grid'
+export { Grid } from './components/Grid'
+export { defaultGridProps } from './components/Grid/constants'
 export { ChartTooltip, DefaultTooltip, ChartTooltipTitle, ChartTooltipValue, ChartTooltipFooter } from './components/ChartTooltip'
 export { EmptyState } from './components/EmptyState'
 export type { EmptyStateProps } from './components/EmptyState'
